+++ specs/specs.cpp
@@ -73,6 +73,7 @@ struct DebugState {
   bool EnableBinBorders = false;
   bool EnablePacifistOverlay = false;
   bool DrawEnemyDetection = false;
+  bool BlackMarketTrainer = false;
   bool IncludeHitboxOrigins = false;
 
   EnabledEntities Ids;
@@ -867,6 +868,132 @@ void drawOverlayWindow() {
 
         probe_idx++;
       } while (probe_idx < 8);
+    }
+  }
+
+  if (gDebugState.BlackMarketTrainer && gGlobalState->screen_state == 0 &&
+      gGlobalState->play_state == 0) {
+    if (gGlobalState->level > 4 && gGlobalState->level < 9 &&
+        gGlobalState->is_worm == 0 && gGlobalState->is_blackmarket == 0) {
+
+      auto color = ImGui::GetColorU32({255.f, 0.0f, 0.0f, 0.25f});
+      auto bm_color = ImGui::GetColorU32({0.f, 255.0f, 0.0f, 0.25f});
+      for (auto idx = 229; idx < 4692; idx++) {
+        auto floor = gGlobalState->level_state->entity_floors[idx];
+        if (!floor) {
+          continue;
+        }
+        auto screen = gameToScreen({floor->x - 0.3f, floor->y + 0.3f});
+
+        if (floor->x == gGlobalState->level_state->alt_exit_x &&
+            floor->y == gGlobalState->level_state->alt_exit_y) {
+          drawEntityHitbox(floor, bm_color, true);
+          continue;
+        }
+
+        if (idx < 229) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Too\nHigh");
+          continue;
+        }
+
+        // Only check specific floor types
+        if (floor->entity_type != 0x2389 && floor->entity_type != 0x2387) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Wrong\nType");
+          continue;
+        }
+
+        if (floor->flag_4 || floor->flag_5 || floor->shopkeeper_tile ||
+            floor->under_door) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Wrong\nFlags");
+          continue;
+        }
+
+        // Floor Below
+        auto floor_below = gGlobalState->level_state->entity_floors[idx + 46];
+        if (!floor_below) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "None\nBelow");
+          continue;
+        }
+
+        if (floor_below->entity_type != 0x2389 &&
+            floor_below->entity_type != 0x2387) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Below\nWrong\nType");
+          continue;
+        }
+
+        if (floor_below->flag_4 || floor_below->flag_5 ||
+            floor_below->shopkeeper_tile || floor_below->under_door) {
+          gOverlayDrawList->AddText(ImGui::GetFont(),
+                                    ImGui::GetFontSize() + 2.f,
+                                    ImVec2{screen.x, screen.y}, IM_COL32_WHITE,
+                                    "Below\nWrong\nFlags");
+          continue;
+        }
+
+        // Floor Left
+        if (idx > 1) {
+          auto floor_left = gGlobalState->level_state->entity_floors[idx - 1];
+          // Left is water
+          if (floor_left && floor_left->entity_type == 0x1b) {
+            gOverlayDrawList->AddText(
+                ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+                ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Left\nWater");
+            continue;
+          }
+        }
+
+        // Floor Right
+        if (idx <= 4690) {
+          auto floor_right = gGlobalState->level_state->entity_floors[idx + 1];
+          // Right is water
+          if (floor_right && floor_right->entity_type == 0x1b) {
+            gOverlayDrawList->AddText(
+                ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+                ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Right\nWater");
+            continue;
+          }
+        }
+
+        // Floor Above
+        if (idx < 46) {
+          auto floor_above = gGlobalState->level_state->entity_floors[idx - 46];
+          // Above is water
+          if (floor_above && floor_above->entity_type == 0x1b) {
+            gOverlayDrawList->AddText(
+                ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+                ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Above\nWater");
+            continue;
+          }
+        }
+
+        // Floor Below Water Check. should be impossible?
+        if (floor_below->entity_type == 0x1b) {
+          gOverlayDrawList->AddText(
+              ImGui::GetFont(), ImGui::GetFontSize() + 2.f,
+              ImVec2{screen.x, screen.y}, IM_COL32_WHITE, "Below\nWater");
+          continue;
+        }
+
+        if (gGlobalState->rushing_water && floor_below->y <= 72) {
+          gOverlayDrawList->AddText(ImGui::GetFont(),
+                                    ImGui::GetFontSize() + 2.f,
+                                    ImVec2{screen.x, screen.y}, IM_COL32_WHITE,
+                                    "Rushing\nWater\nCutoff");
+          continue;
+        }
+
+        drawEntityHitbox(floor, color, true);
+      }
     }
   }
 
@@ -2005,6 +2132,7 @@ void drawDebugTab() {
   ImGui::Checkbox("Draw Bin Borders", &gDebugState.EnableBinBorders);
   ImGui::Checkbox("Draw Owned Entities", &gDebugState.EnablePacifistOverlay);
   ImGui::Checkbox("Draw Detection Boxes", &gDebugState.DrawEnemyDetection);
+  ImGui::Checkbox("Black Market Trainer", &gDebugState.BlackMarketTrainer);
   ImGui::Checkbox("Include Hitbox Origins", &gDebugState.IncludeHitboxOrigins);
   ImGui::Checkbox("Include Floor Decorations", &gDebugState.IncludeFloorDecos);
   if (ImGui::Checkbox("Disable Olmec Spawns",
@@ -2245,7 +2373,7 @@ void drawDebugTab() {
         ImGui::Text("%04d: %s", ent->entity_type, ent->TypeName());
 
         ImGui::TableNextColumn();
-        ImGui::Text("%f", ent->current_z);
+        ImGui::Text("%.4f", ent->current_z);
 
         ImGui::TableNextColumn();
         ImGui::Text("%d", ent->flag_deletion);
@@ -2299,9 +2427,9 @@ void drawSelectedEntityTab() {
     ImGui::SliderFloat("height", &gSelectedEntityState.Entity->height, 0.0,
                        10.0);
     ImGui::SliderFloat("current_z", &gSelectedEntityState.Entity->current_z,
-                       0.0, 50.0, "%.6f");
+                       0.0, 50.0, "%.4f");
     ImGui::SliderFloat("original_z", &gSelectedEntityState.Entity->original_z,
-                       0.0, 50.0, "%.6f");
+                       0.0, 50.0, "%.4f");
     ImGui::SliderFloat("alpha", &gSelectedEntityState.Entity->alpha, 0.0, 1.0);
     ImGui::SliderFloat("hitbox up", &gSelectedEntityState.Entity->hitbox_up,
                        0.0, 5.0);